model/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2

# inference deps
torch
transformers>=4.45.0
accelerate
pillow
opencv-python
timm
einops
sentencepiece
